refactor(test): extract shared DOMMatrix and DOMPoint polyfills for JSDOM. - Created __tests__/setup/dom-polyfills.ts to house shared SVG-related polyfills. - Replaced redundant inline implementations in coordinate.test.ts and zoom-wheel.test.ts. - Improved consistency and maintainability of the test suite.
