+++ __tests__/unit/editor/utils/coordinate.test.ts
@@ -7,6 +7,7 @@ import {
   viewportToClient,
 } from '../../../../src/editor/utils/coordinate';
 import type { Element } from '../../../../src/types';
+import '../../../setup/dom-polyfills';
 
 const createSVG = (matrix?: DOMMatrix) => {
   const svg = document.createElementNS('http://www.w3.org/2000/svg', 'svg');
@@ -66,34 +67,3 @@ describe('editor/utils/coordinate', () => {
     expect(bounds.height).toBe(60);
   });
 });
-// Polyfill DOMMatrix/DOMPoint for jsdom
-class Matrix {
-  e = 0;
-  f = 0;
-  translate(x = 0, y = 0) {
-    this.e += x;
-    this.f += y;
-    return this;
-  }
-  inverse() {
-    const m = new Matrix();
-    m.e = -this.e;
-    m.f = -this.f;
-    return m;
-  }
-}
-
-class Point {
-  constructor(
-    public x: number,
-    public y: number,
-  ) {}
-  matrixTransform(matrix: any) {
-    return new Point(this.x + (matrix?.e || 0), this.y + (matrix?.f || 0));
-  }
-}
-
-// @ts-expect-error test polyfill
-globalThis.DOMMatrix ??= Matrix;
-// @ts-expect-error test polyfill
-globalThis.DOMPoint ??= Point;
